requirements.txt
sounddevice
scipy
numpy
pillow
keyboard
tkinter